# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kampus Ariza ve Talep Sistemi API",
        "description": "Universite kampusu icin destek talep yonetim API dokumantasyonu",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/priorities": {
            "get": {
                "tags": [
                    "Priorities"
                ],
                "summary": "Öncelikleri listele",
                "operationId": "2d18ab468322b1bf3f5181b349b181cd",
                "responses": {
                    "200": {
                        "description": "Liste döndü"
                    }
                }
            },
            "post": {
                "tags": [
                    "Priorities"
                ],
                "summary": "Yeni öncelik oluştur",
                "operationId": "be2d141500ff477153ade637ed36c3d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oluşturuldu"
                    }
                }
            }
        },
        "/api/priorities/{priority}": {
            "get": {
                "tags": [
                    "Priorities"
                ],
                "summary": "Öncelik detayı",
                "operationId": "97f293a9516000db66222af5f69dab16",
                "parameters": [
                    {
                        "name": "priority",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detay döndü"
                    }
                }
            },
            "put": {
                "tags": [
                    "Priorities"
                ],
                "summary": "Öncelik güncelle",
                "operationId": "ae0fc12d9cf37b360900d64500caa4f1",
                "parameters": [
                    {
                        "name": "priority",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Güncellendi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Priorities"
                ],
                "summary": "Öncelik sil",
                "operationId": "c47d818da2912ca9e96ff62483c43d7c",
                "parameters": [
                    {
                        "name": "priority",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Silindi"
                    }
                }
            }
        },
...
    },
...
}